R56

+++ app/database/migrations/2021_09_03_180704_add_spy_and_wizard_strength_costs_to_units.php
@@ -14,8 +14,8 @@ class AddSpyAndWizardStrengthCostsToUnits extends Migration
     public function up()
     {
         Schema::table('units', function (Blueprint $table) {
-            $table->integer('cost_spy_strength')->default(0)->after('cost_morale');
-            $table->integer('cost_wizard_strength')->default(0)->after('cost_spy_strength');
+            $table->decimal('cost_spy_strength')->default(0)->after('cost_morale');
+            $table->decimal('cost_wizard_strength')->default(0)->after('cost_spy_strength');
         });
     }
